Move components extracted from MyOrders to `/react/components/legacy`
