+++ react/components/legacy/BundleItem.js
