+++ react/components/legacy/FormattedPrice.js
