+++ react/components/legacy/PaymentConnectorResponses.js
