+++ react/components/legacy/Product.js
