+++ react/components/legacy/Total.js
