+++ react/components/legacy/Totals.js
